+++ src/kiara/workflow.py
@@ -77,6 +77,10 @@ class KiaraWorkflow(object):
     def inputs(self) -> ValueSet:
         return self.pipeline.inputs
 
+    @property
+    def outputs(self) -> ValueSet:
+        return self.pipeline.outputs
+
     def get_current_state(self) -> PipelineState:
         return self.pipeline.get_current_state()
 
@@ -89,16 +93,12 @@ class KiaraWorkflow(object):
     #     self.pipeline.set_pipeline_inputs(**inputs)
 
     @property
-    def outputs(self) -> ValueSet:
-        return self.pipeline.outputs
-
-    @property
     def input_names(self) -> typing.List[str]:
-        return list(self.inputs.keys())
+        return list(self.inputs.get_all_field_names())
 
     @property
     def output_names(self) -> typing.List[str]:
-        return list(self.outputs.keys())
+        return list(self.outputs.get_all_field_names())
 
     @property
     def workflow_id(self) -> str: